+++ lib/libcalendar.dart
@@ -1,7 +1,1 @@
-///
-/// This package contains some utility classes and method for manipulating
-/// calendar and date time.
-///
-library libcalendar;
-
 export 'src/calendar_converter.dart';