+++ test/cjdn_islamic.dart
@@ -2,29 +2,29 @@ import 'package:libcalendar/libcalendar.dart';
 import 'package:test/test.dart';
 
 void main() {
-  group('fromIslamicToCJDN() method tests.', () {
-    fromIslamicToCJDNYearParameterLessThan1317();
-    fromIslamicToCJDNYearParameterMoreThan1524();
-    fromIslamicToCJDNMonthParameterLessThan1();
-    fromIslamicToCJDNMonthParameterMoreThan12();
-    fromIslamicToCJDNDayParameterLessThan1();
-    fromIslamicToCJDNDayParameterMoreThan30();
-    fromIslamicToCJDNDateBeforeShaaban28th1317();
-    fromIslamicToCJDNDateAfterShawwal29th1524();
-    fromIslamicToCJDNConvertShaaban29th1432ToCJDN();
+  group('fromIslamicToCjdn() method tests.', () {
+    fromIslamicToCjdnYearParameterLessThan1317();
+    fromIslamicToCjdnYearParameterMoreThan1524();
+    fromIslamicToCjdnMonthParameterLessThan1();
+    fromIslamicToCjdnMonthParameterMoreThan12();
+    fromIslamicToCjdnDayParameterLessThan1();
+    fromIslamicToCjdnDayParameterMoreThan30();
+    fromIslamicToCjdnDateBeforeShaaban28th1317();
+    fromIslamicToCjdnDateAfterShawwal29th1524();
+    fromIslamicToCjdnConvertShaaban29th1432ToCJDN();
   });
 
-  group('fromCJDNtoIslamic() method tests.', () {
-    fromCJDNtoIslamicCjdnParameterLessThan2415021();
-    fromCJDNtoIslamicCjdnParameterMoreThan2488434();
-    fromCJDNtoIslamicConvert2455774ToGregorian();
+  group('fromCjdnToIslamic() method tests.', () {
+    fromCjdnToIslamicCjdnParameterLessThan2415021();
+    fromCjdnToIslamicCjdnParameterMoreThan2488434();
+    fromCjdnToIslamicConvert2455774ToGregorian();
   });
 }
 
-void fromIslamicToCJDNYearParameterLessThan1317() {
+void fromIslamicToCjdnYearParameterLessThan1317() {
   test('Invalid year parameter (less than 1317) should throw error.', () {
     expect(
-      () => fromIslamicToCJDN(1316, 1, 1),
+      () => fromIslamicToCjdn(1316, 1, 1),
       throwsA(
         predicate(
           (e) =>
@@ -36,10 +36,10 @@ void fromIslamicToCJDNYearParameterLessThan1317() {
   });
 }
 
-void fromIslamicToCJDNYearParameterMoreThan1524() {
+void fromIslamicToCjdnYearParameterMoreThan1524() {
   test('Invalid year parameter (more than 1524) should throw error.', () {
     expect(
-      () => fromIslamicToCJDN(1525, 1, 1),
+      () => fromIslamicToCjdn(1525, 1, 1),
       throwsA(
         predicate(
           (e) =>
@@ -51,10 +51,10 @@ void fromIslamicToCJDNYearParameterMoreThan1524() {
   });
 }
 
-void fromIslamicToCJDNMonthParameterLessThan1() {
+void fromIslamicToCjdnMonthParameterLessThan1() {
   test('Invalid month parameter (less than 1) should throw error.', () {
     expect(
-      () => fromIslamicToCJDN(1430, 0, 1),
+      () => fromIslamicToCjdn(1430, 0, 1),
       throwsA(
         predicate(
           (e) =>
@@ -66,10 +66,10 @@ void fromIslamicToCJDNMonthParameterLessThan1() {
   });
 }
 
-void fromIslamicToCJDNMonthParameterMoreThan12() {
+void fromIslamicToCjdnMonthParameterMoreThan12() {
   test('Invalid month parameter (more than 12) should throw error.', () {
     expect(
-      () => fromIslamicToCJDN(1430, 13, 1),
+      () => fromIslamicToCjdn(1430, 13, 1),
       throwsA(
         predicate(
           (e) =>
@@ -81,10 +81,10 @@ void fromIslamicToCJDNMonthParameterMoreThan12() {
   });
 }
 
-void fromIslamicToCJDNDayParameterLessThan1() {
+void fromIslamicToCjdnDayParameterLessThan1() {
   test('Invalid day parameter (less than 1) should throw error.', () {
     expect(
-      () => fromIslamicToCJDN(1430, 1, 0),
+      () => fromIslamicToCjdn(1430, 1, 0),
       throwsA(
         predicate(
           (e) =>
@@ -96,10 +96,10 @@ void fromIslamicToCJDNDayParameterLessThan1() {
   });
 }
 
-void fromIslamicToCJDNDayParameterMoreThan30() {
+void fromIslamicToCjdnDayParameterMoreThan30() {
   test('Invalid day parameter (more than 30) should throw error.', () {
     expect(
-      () => fromIslamicToCJDN(1430, 1, 31),
+      () => fromIslamicToCjdn(1430, 1, 31),
       throwsA(
         predicate(
           (e) =>
@@ -111,11 +111,11 @@ void fromIslamicToCJDNDayParameterMoreThan30() {
   });
 }
 
-void fromIslamicToCJDNDateBeforeShaaban28th1317() {
+void fromIslamicToCjdnDateBeforeShaaban28th1317() {
   test('Invalid date range (before Sha\'aban 28th, 1317) should throw error.',
       () {
     expect(
-      () => fromIslamicToCJDN(1317, 8, 27),
+      () => fromIslamicToCjdn(1317, 8, 27),
       throwsA(
         predicate(
           (e) =>
@@ -129,10 +129,10 @@ void fromIslamicToCJDNDateBeforeShaaban28th1317() {
   });
 }
 
-void fromIslamicToCJDNDateAfterShawwal29th1524() {
+void fromIslamicToCjdnDateAfterShawwal29th1524() {
   test('Invalid date range (after Shawwal 29th, 1524) should throw error.', () {
     expect(
-      () => fromIslamicToCJDN(1524, 11, 1),
+      () => fromIslamicToCjdn(1524, 11, 1),
       throwsA(
         predicate(
           (e) =>
@@ -146,16 +146,16 @@ void fromIslamicToCJDNDateAfterShawwal29th1524() {
   });
 }
 
-void fromIslamicToCJDNConvertShaaban29th1432ToCJDN() {
+void fromIslamicToCjdnConvertShaaban29th1432ToCJDN() {
   test('Convert Sha\'aban 29th, 1432 to CJDN.', () {
-    expect(fromIslamicToCJDN(1432, 8, 29), equals(2455774));
+    expect(fromIslamicToCjdn(1432, 8, 29), equals(2455774));
   });
 }
 
-void fromCJDNtoIslamicCjdnParameterLessThan2415021() {
+void fromCjdnToIslamicCjdnParameterLessThan2415021() {
   test('Invalid cjdn parameter (less than 2415021) should throw error.', () {
     expect(
-      () => fromCJDNtoIslamic(2415020),
+      () => fromCjdnToIslamic(2415020),
       throwsA(
         predicate(
           (e) =>
@@ -168,10 +168,10 @@ void fromCJDNtoIslamicCjdnParameterLessThan2415021() {
   });
 }
 
-void fromCJDNtoIslamicCjdnParameterMoreThan2488434() {
+void fromCjdnToIslamicCjdnParameterMoreThan2488434() {
   test('Invalid cjdn parameter (more than 2488434) should throw error.', () {
     expect(
-      () => fromCJDNtoIslamic(2488435),
+      () => fromCjdnToIslamic(2488435),
       throwsA(
         predicate(
           (e) =>
@@ -184,8 +184,8 @@ void fromCJDNtoIslamicCjdnParameterMoreThan2488434() {
   });
 }
 
-void fromCJDNtoIslamicConvert2455774ToGregorian() {
+void fromCjdnToIslamicConvert2455774ToGregorian() {
   test('Convert CJDN 2455774 to Islamic.', () {
-    expect(fromCJDNtoIslamic(2455774), equals(new DateTime.utc(1432, 8, 29)));
+    expect(fromCjdnToIslamic(2455774), equals(DateTime.utc(1432, 8, 29)));
   });
 }