+++ CHANGELOG.md
@@ -19,3 +19,13 @@
 
 - **[BREAKING CHANGES]**: Updated Dart SDK requirement.
 - **[BREAKING CHANGES]**: Upgraded all dependencies to latest version (as on October 4th, 2018).
+
+## 0.3.0
+
+- **[BREAKING CHANGES]**: `libcalendar` now requires Dart 3 SDK.
+- **[BREAKING CHANGES]**: Upgraded all dependencies to latest version (as on January 6th, 2025).
+- **[BREAKING CHANGES]**: Renamed `fromGregorianToCJDN` to `fromGregorianToCjdn`.
+- **[BREAKING CHANGES]**: Renamed `fromCJDNtoGregorian` to `fromCjdnToGregorian`.
+- **[BREAKING CHANGES]**: Renamed `fromIslamicToCJDN` to `fromIslamicToCjdn`.
+- **[BREAKING CHANGES]**: Renamed `fromCJDNtoIslamic` to `fromCjdnToIslamic`.
+- `CalendarConversionError` now public.